Implement basic control logic
