+++ midi_light_show/src/lib.rs
@@ -5,18 +5,13 @@ extern crate log;
 extern crate midir;
 extern crate rimd;
 
-mod controls;
-mod song;
+pub mod controls;
 pub mod light;
+mod song;
 
 use log::{error, info, warn};
-use midir::MidiOutput;
 use rimd::{SMFError, TrackEvent, SMF};
-use std::collections::HashSet;
-use std::error::Error;
 use std::path::Path;
-use std::thread::sleep;
-use std::time::Duration;
 
 pub const DEFAULT_VEC_CAPACITY: usize = 133000;
 
@@ -131,7 +126,7 @@ impl MidiTimeInfo {
 #[derive(Copy, Clone)]
 pub struct DeltaTiming(pub i16);
 
-pub fn load_midi_file(pathstr: &str) -> (Vec<TrackEvent>, DeltaTiming) {
+fn load_midi_file(pathstr: &str) -> (Vec<TrackEvent>, DeltaTiming) {
     let mut events: Vec<TrackEvent> = Vec::with_capacity(DEFAULT_VEC_CAPACITY);
 
     let mut division: i16 = 0;
@@ -215,89 +210,6 @@ pub fn transform_events(track_events: Vec<TrackEvent>) -> Vec<MidiEvent> {
     }
 
     events
-}
-
-/// Process all of the MIDI events and send them to output_device
-/// Each note calls `thread::sleep` based on its `vtime` attribute
-/// and `delta_timing`
-pub fn play_from_beginning(
-    output_device: usize,
-    midi_events: Vec<MidiEvent>,
-    delta_timing: DeltaTiming,
-    midi_sender: channel::Sender<NoteEvent>,
-) -> Result<(), Box<dyn Error>> {
-    let midi_out = MidiOutput::new("MIDI Magic Machine")?;
-
-    let port_number = &midi_out.ports()[output_device];
-    let mut conn_out = midi_out.connect(port_number, "led_midi_show")?;
-
-    const DEFAULT_MICROS_PER_QNOTE: u64 = 681817;
-    let mut micros_per_tick = (DEFAULT_MICROS_PER_QNOTE as f32 / delta_timing.0 as f32) as u64;
-
-    println!("[ [   Show Starts Now   ] ]");
-    {
-        // Define a new scope in which the closure `sleep_play` borrows conn_out, so it can be called easily
-        let mut sleep_play = |midi: &MidiEvent| match midi {
-            MidiEvent::Tempo(tempo_change) => {
-                let u = (tempo_change.micros_per_qnote as f32 / delta_timing.0 as f32) as u64;
-                info!("Update micros per tick: {}", u);
-                micros_per_tick = u;
-            }
-            MidiEvent::Note(note) => {
-                sleep(Duration::from_micros(note.vtime * micros_per_tick));
-
-                if let Err(e) = midi_sender.send(note.clone()) {
-                    error!("send err {:?}", e)
-                }
-
-                let _ = match note.channel_event {
-                    ChannelEvent::ChannelOn(c) => conn_out.send(&[c, note.note, note.velocity]),
-                    ChannelEvent::ChannelOff(c) => conn_out.send(&[c, note.note, note.velocity]),
-                };
-            }
-            MidiEvent::SustainPedal(p) => info!("Sustain pedal: {:?}", p),
-        };
-
-        let mut pedal_state = PedalState::Off;
-        let mut sustained = HashSet::new();
-
-        for n in midi_events {
-            match (&n, pedal_state) {
-                (MidiEvent::SustainPedal(SustainPedalEvent(PedalState::Off)), PedalState::On) => {
-                    for s in sustained.drain() {
-                        sleep_play(&MidiEvent::Note(s));
-                    }
-                    pedal_state = PedalState::Off;
-                }
-                (MidiEvent::SustainPedal(SustainPedalEvent(PedalState::On)), PedalState::Off) => {
-                    pedal_state = PedalState::On;
-                }
-                (
-                    MidiEvent::Note(NoteEvent {
-                        channel_event: ChannelEvent::ChannelOff(c),
-                        time,
-                        vtime,
-                        note,
-                        velocity,
-                    }),
-                    PedalState::On,
-                ) => {
-                    sustained.insert(NoteEvent {
-                        channel_event: ChannelEvent::ChannelOff(*c),
-                        time: *time,
-                        vtime: *vtime,
-                        note: *note,
-                        velocity: *velocity,
-                    });
-                }
-                (n, _p) => sleep_play(&n),
-            }
-        }
-    }
-
-    // This is optional, the connection would automatically be closed as soon as it goes out of scope
-    conn_out.close();
-    Ok(())
 }
 
 // figure out how many microsec are in a quarter note
